+++ src/Service/InputFilter/ListAddressesFactory.php
@@ -14,11 +14,11 @@ use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
 /**
- * Input filter factory for address balance request
+ * Input filter factory for list addresses request
  *
- * Creates input filter for address balance request validation
+ * Creates input filter for list addresses request validation
  */
-class AddressBalanceFactory implements FactoryInterface
+class ListAddressesFactory implements FactoryInterface
 {
     /**
      * Create service
@@ -30,22 +30,9 @@ class AddressBalanceFactory implements FactoryInterface
     {
         $factory = new Factory();
         $inputFilter = $factory->createInputFilter(array(
-            'address' => array(
-                'name'       => 'address',
-                'required'   => true,
-                'validators' => array(
-                    array(
-                        'name' => 'not_empty',
-                        'break_chain_on_failure' => true,
-                    ),
-                    array(
-                        'name' => '\Sake\BlockchainWalletApi\Validator\BitcoinAddress',
-                    ),
-                ),
-            ),
             'confirmations' => array(
                 'name'       => 'confirmations',
-                'required'   => true,
+                'required'   => false,
                 'validators' => array(
                     array(
                         'name' => 'greater_than',