+++ src/Service/InputFilter/WalletBalanceFactory.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Sake
+ *
+ * @link      http://github.com/sandrokeil/BlockchainWalletApi for the canonical source repository
+ * @copyright Copyright (c) 2014 Sandro Keil
+ * @license   http://github.com/sandrokeil/BlockchainWalletApi/blob/master/LICENSE.txt New BSD License
+ */
+
+namespace Sake\BlockchainWalletApi\Service\InputFilter;
+
+use Zend\InputFilter\InputFilter;
+use Zend\ServiceManager\FactoryInterface;
+use Zend\ServiceManager\ServiceLocatorInterface;
+
+/**
+ * Input filter factory for wallet balance request
+ *
+ * Creates input filter for wallet balance request validation
+ */
+class WalletBalanceFactory implements FactoryInterface
+{
+    /**
+     * Create service
+     *
+     * @param ServiceLocatorInterface $serviceLocator
+     * @return mixed
+     */
+    public function createService(ServiceLocatorInterface $serviceLocator)
+    {
+        return new InputFilter();
+    }
+}