+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # BlockchainWalletApi CHANGELOG
 
+## Version 1.0.0 - TBA
 
+* Validator support via input filter
+* Service plugin manager for responses, requests and input filter
+* Hydrator strategies for address list and addresses
+* Response/Request classes for all blockchain wallet api methods